PlutoStaticHTML

Convert Pluto notebooks to pure HTML (without Javascript). Small discussion with Fons van der Plas at https://github.com/fonsp/Pluto.jl/discussions/1607.

I'm using this package for my own blog, for example: https://huijzer.xyz/posts/frequentist-bayesian-coin-flipping/. A link to the notebook is at the bottom of the page.

The most important methods are notebook2html or parallel_build (API).

Note

notebook2html and parallel_build evaluate notebooks after copying your notebooks. This ensures that the original notebook will not be changed. The copied notebook will be placed in the same folder so that @__DIR__ can be used to locate files relative to the notebook path. Add **/_tmp_* to your .gitignore to ignore the temporary copies.

notebook2html

To process one notebook. To make use of caching or parallel building, use parallel_build.

Example usage for notebook2html:

julia> using PlutoStaticHTML

julia> notebook2html("Exciting analysis.jl") |> print
<div class="markdown"><p>This is an example notebook.</p>
[...]

Note that, in general, what works best is to

Determine a list of paths to your notebooks.

Pass the paths to parallel_build which writes the HTML outputs to files by default.

Read the output from the HTML files and show it inside a static website.

More specific instructions for

Documenter.jl

Franklin.jl

Parallel build

are listed below.

Documenter.jl

To see how to embed output from a Pluto notebook in Documenter.jl, checkout "make.jl" in this repository.

Warn

It is typically not a good idea to call the conversion from inside Documenter.jl. For some reason, that is likely to freeze or hang; probably due to stdout being flooded with information or something.

Franklin.jl

See the next section for a parallel version. In utils.jl define:

"""
    lx_pluto(com, _)

Embed a Pluto notebook via:
https://github.com/rikhuijzer/PlutoStaticHTML.jl
"""
function lx_pluto(com, _)
    file = string(Franklin.content(com.braces[1]))::String
    notebook_path = joinpath("posts", "notebooks", "$file.jl")
    log_path = joinpath("posts", "notebooks", "$file.log")

    return """
        ```julia:pluto
        # hideall

        using PlutoStaticHTML: notebook2html

        path = "$notebook_path"
        log_path = "$log_path"
        @assert isfile(path)
        @info "→ evaluating Pluto notebook at (\$path)"
        html = open(log_path, "w") do io
            redirect_stdout(io) do
                html = notebook2html(path)
                return html
            end
        end
        println("~~~\n\$html\n~~~\n")
        ```
        \\textoutput{pluto}
        """
end

Next, the Pluto notebook at "/posts/notebooks/analysis.jl" can be included in a Franklin webpage. For example:

+++
title = "My analysis"
+++

\pluto{analysis}

Parallel build

The approach above lets Franklin.jl handle the build. This doesn't work in parallel. To speed up the build, this package defines parallel_build. parallel_build evaluates the notebooks in parallel by default. Also, it can use Caching to speed up the build even more.

To use parallel_build, pass a dir to write HTML files for all notebook files (recognized by ".jl" extension):

julia> using PlutoStaticHTML: parallel_build

julia> dir = joinpath("posts", "notebooks");

julia> parallel_build(BuildOptions(dir));

To run only specific notebooks, use:

julia> files = ["notebook1.jl", "notebook2.jl"];

julia> parallel_build(BuildOptions(dir), files)

In CI, be sure to call this before using Franklin serve or optimize. Next, read the HTML back into Franklin by defining in utils.jl:

"""
    lx_readhtml(com, _)

Embed a Pluto notebook via:
https://github.com/rikhuijzer/PlutoStaticHTML.jl
"""
function lx_readhtml(com, _)
    file = string(Franklin.content(com.braces[1]))::String
    dir = joinpath("posts", "notebooks")
    filename = "$(file).jl"

    return """
        ```julia:pluto
        # hideall

        filename = "$filename"
        html = read(filename, String)
        println("~~~\n\$html\n~~~\n")
        ```
        \\textoutput{pluto}
        """
end

and calling this from Franklin. For example:

+++
title = "My analysis"
reeval = true
+++

\readhtml{analysis}

Caching

Using caching can greatly speed up running times by avoiding to re-evaluate notebooks. Caching can be enabled by passing previous_dir via BuildOptions. This previous_dir should point to a location where HTML files are from the previous build. Then, parallel_build will, for each input file file.jl, check:

Whether joinpath(previous_dir, "file.html") exists

Whether the SHA checksum of the current $file.jl matches the checksum of the previous $file.jl. When assuming that Pluto's built-in package manager is used to manage packages, this check ensures that the packages of the previous run match the packages of the current run.

Whether the Julia version of the previous run matches the Julia version of the current run.

Note

Caching can only be used if the notebooks are deterministic, that is, the notebook will always produce the same output from the same input.

Note

The previous_dir provides a lot of flexibility. For example, it is possible to point towards a Git directory with the HTML output files from last time. Alternatively, it is possible to download the web pages where the notebooks are shown and put these web pages in a directory. This works because this package extracts the state from the previous run automatically.

LaTeX equations

With Franklin.jl, update foot_katex.html to include:

<script>
  const options = {
    delimiters: [
      {left: "$$", right: "$$", display: true},
      {left: "$", right: "$", display: false},
      {left: "\\begin{equation}", right: "\\end{equation}", display: true},
      {left: "\\begin{align}", right: "\\end{align}", display: true},
      {left: "\\begin{alignat}", right: "\\end{alignat}", display: true},
      {left: "\\begin{gather}", right: "\\end{gather}", display: true},
      {left: "\\(", right: "\\)", display: false},
      {left: "\\[", right: "\\]", display: true}
    ]
  };
  renderMathInElement(document.body, options);
</script>

which basically ensures that inline math surrounded by single dollar symbols is also rendered. Note that Pluto.jl runs MathJax by default which might sometimes cause inconsistencies between the math in Pluto and inside your HTML.

API

PlutoStaticHTML.HTMLOptions — Type

HTMLOptions(;
    code_class::AbstractString="language-julia",
    output_class::AbstractString="code-output",
    output_pre_class::AbstractString="documenter-example-output",
    hide_code::Bool=false,
    hide_md_code::Bool=true,
    hide_md_def_code::Bool=true,
    add_state::Bool=true,
    append_build_context::Bool=false,
    compiler_options::Union{Nothing,CompilerOptions}=nothing
)

Options for notebook2html:

code_class: HTML class for code. This is used by CSS and/or the syntax highlighter.

outputpreclass: HTML class for<pre>`.

output_class: HTML class for output. This is used by CSS and/or the syntax highlighter.

hide_code: Whether to omit all code blocks. Can be useful when readers are not interested in code at all.

hide_md_code: Whether to omit all Markdown code blocks.

hide_md_def_code: Whether to omit Franklin Markdown definition code blocks (blocks surrounded by +++).

add_state: Whether to add a comment in HTML with the state of the input notebook. This state can be used for caching. Specifically, this state stores a checksum of the input notebook and the Julia version.

append_build_context: Whether to append build context. When set to true, this adds information about the dependencies and Julia version. This is not executed via Pluto.jl's evaluation to avoid having to add extra dependencies to existing notebooks. Instead, this reads the manifest from the notebook file.

compiler_options: Pluto.Configuration.CompilerOptions to be passed to Pluto. This can, for example, be useful to pass custom system images from PackageCompiler.jl.

PlutoStaticHTML.BuildOptions — Type

BuildOptions(
    dir::AbstractString;
    write_files::Bool=true,
    previous_dir::Union{Nothing,AbstractString}=nothing,
    output_format::OutputFormat=html_output,
    use_distributed::Bool=true
)

Options for parallel_build:

dir: Directory in which the Pluto notebooks are stored.

write_files: Write files to joinpath(dir, "$file.html").

previous_dir::Union{Nothing,AbstractString}=Nothing: Use the output from the previous run as a cache to speed up running time. To use the cache, specify a directory previous_dir::AbstractString which contains HTML files from a previous run. Specifically, files are expected to be at joinpath(previous_dir, "$file.html"). The output from the previous run may be embedded in a larger HTML web page. This package will extract the original output from the full HTML web page. By default, caching is disabled.

output_format: What file to write the output to. By default this is html_output::OutputFormat meaning that the output of the HTML method is pure HTML. To generate Franklin or files, use franklin_output. When BuildOptions.write_files == true and output_format != html_output, the output file has a ".md" extension instead of ".html".

use_distributed: Whether to build the notebooks in different processes. By default, this is enabled just like in Pluto and the notebooks are build in parallel. The benefit of different processes is that things are more independent of each other. Unfortunately, the drawback is that compilation has to happen for each process. By setting this option to false, all notebooks are built sequentially in the same process which avoids recompilation. This is likely quicker in situations where there are few threads available such as GitHub Runners depending on the notebook contents. Beware that use_distributed=false will not work with Pluto's built-in package manager.

PlutoStaticHTML.parallel_build — Function

parallel_build(
    bopts::BuildOptions,
    files,
    hopts::HTMLOptions=HTMLOptions();
    session=ServerSession()
) -> Vector{String}

Build all files in dir in parallel.

parallel_build(
    bopts::BuildOptions,
    hopts::HTMLOptions=HTMLOptions()
) -> Vector{String}

Build all ".jl" files in dir in parallel.

PlutoStaticHTML.notebook2html — Function

notebook2html(nb::Notebook, path, opts::HTMLOptions=HTMLOptions()) -> String

Return the code and output as HTML for nb. Assumes that the notebook has already been executed.

notebook2html(
    path::AbstractString,
    opts::HTMLOptions=HTMLOptions();
    session=ServerSession(),
    append_cells=Cell[],
) -> String

Run the Pluto notebook at path and return the code and output as HTML. This makes a copy of the notebook at path and runs it.

Keyword arguments:

append_cells: Specify one or more Pluto.Cells to be appended at the end of the notebook. Be careful when adding new packages via this method because it may disable Pluto.jl's built-in package management.